Add OpenAI API key support and refactor wallet logic
