+++ examples/xmtp-group-toss/.env.example
@@ -3,6 +3,7 @@ ENCRYPTION_KEY= # the encryption key for the wallet
 # public key is 
 
 NETWORK_ID=base-sepolia # base-mainnet or others
+OPENAI_API_KEY= # the OpenAI API key  
 CDP_API_KEY_NAME= # the name of the CDP API key
 CDP_API_KEY_PRIVATE_KEY= # the private key for the CDP API key
 XMTP_ENV=dev # local, dev, production
