+++ examples/xmtp-group-toss/package.json
@@ -6,10 +6,10 @@
   "scripts": {
     "build": "tsc",
     "clean": "cd ../../ && rm -rf examples/xmtp-group-toss/.data",
-    "dev": "tsx --watch index.ts",
+    "dev": "tsx --watch src/index.ts",
     "gen:keys": "tsx ../../scripts/generateKeys.ts",
     "lint": "cd ../.. && yarn eslint examples/xmtp-group-toss",
-    "start": "tsx index.ts"
+    "start": "tsx src/index.ts"
   },
   "dependencies": {
     "@coinbase/agentkit": "^0.2.0",
@@ -18,8 +18,9 @@
     "@langchain/core": "^0.3.20",
     "@langchain/langgraph": "^0.2.24",
     "@langchain/openai": "^0.3.14",
-    "@xmtp/node-sdk": "1.0.2",
+    "@xmtp/node-sdk": "1.2.0",
     "dotenv": "^16.4.7",
+    "redis": "^4.7.0",
     "uint8arrays": "^5.1.0",
     "viem": "^2.7.15"
   },
