+++ examples/xmtp-group-toss/src/helper.ts
@@ -1,0 +1,27 @@
+// Interface for parsed JSON response
+export interface TossJsonResponse {
+  topic?: string;
+  options?: string[];
+  amount?: string;
+  valid?: boolean;
+  reason?: string;
+}
+/**
+ * Extract JSON from agent response text
+ * @param response The text response from agent
+ * @returns Parsed JSON object or null if not found
+ */
+export function extractJsonFromResponse(
+  response: string,
+): TossJsonResponse | null {
+  try {
+    const jsonMatch = response.match(/\{[\s\S]*\}/);
+    if (jsonMatch) {
+      return JSON.parse(jsonMatch[0]) as TossJsonResponse;
+    }
+    return null;
+  } catch (error) {
+    console.error("Error parsing JSON from agent response:", error);
+    return null;
+  }
+}
